requirements.txt
Flask
requests
werkzeug
gunicorn
python-dotenv
openai